Cast ns to s

+++ src/query_ollama/structs.rs
@@ -1,6 +1,6 @@
 #[derive(Debug)]
 pub struct ValidationResults {
-    pub total_duration: u32,
+    pub total_duration: f32,
     pub reasoning: String,
     pub valid_instructions: bool,
 }
